# openapi/libraryz.yaml
openapi: 3.1.0
info:
  title: LibraryZ API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    LibraryZ backend: auth and sessions, catalog (works + editions with
    content-addressed file storage), crowdsourced edits with moderation,
    personal libraries, and recommendations. See PLAN.md and ARCHITECTURE.md
    in the repo root.

    **Every operation has a cURL sample** (`x-codeSamples`). They assume:

    ```sh
    BASE=http://localhost:8080          # Kong in `docker compose up`
    TOKEN=...      REFRESH=...          # from POST /auth/login (see its sample)
    WORK_ID=...    EDITION_ID=...    CONTRIBUTION_ID=...
    ```

    **Authentication.** `POST /auth/login` returns a short-lived access token
    (JWT, 15 minutes by default) and a single-use refresh token. Send the
    access token as `Authorization: Bearer <token>`. On a `401`, call
    `POST /auth/refresh` once and retry; if that also fails, log in again.
    Presenting an already-used refresh token revokes the whole session.

    **Errors.** `4xx`/`5xx` bodies are short `text/plain` messages (`4xx`
    messages are safe to show to users), except where an operation documents
    a JSON body (e.g. the `409` on edition upload). `5xx` bodies are generic;
    details are only logged server-side.

    **Rate limits.** In the compose/production setup Kong enforces per-IP
    limits (login 5/min, signup 3/min, refresh/logout 20/min, edition upload
...
servers:
...
    description: Local dev
...
security: []
paths:
...
  /works/{id}/contributions:
    post:
      operationId: submitContribution
      tags: [contributions]
      summary: Submit a proposed edit to a work
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContributionRequest"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s -X POST "$BASE/works/$WORK_ID/contributions" -H "Authorization: Bearer $TOKEN" -H 'Content-Type: application/json' \
              -d '{"patch":{"title":"Moby-Dick; or, The Whale","publication_year":1851}}'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contributions
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/contributions?status=pending&limit=50"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contribution" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /contributions/{id}:
    get:
      operationId: getContribution
      tags: [contributions]
      summary: Get a single contribution
      parameters:
        - $ref: "#/components/parameters/UUIDPathId"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/contributions/$CONTRIBUTION_ID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contribution" }
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /me/contributions:
    get:
      operationId: listMyContributions
      tags: [contributions]
      summary: List the calling user's own contributions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -s "$BASE/me/contributions?status=pending" -H "Authorization: Bearer $TOKEN"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contribution" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token from `POST /auth/login` or `POST /auth/refresh`. Expires
        after `LIBRARYZ_ACCESS_TOKEN_TTL` (15m); revoked immediately by
        `POST /auth/logout-all`.
  parameters:
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Bad request
      content:
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: >
        Missing, malformed, expired or revoked access token (refresh it with
        `POST /auth/refresh`), or wrong credentials on login
      content:
        text/plain:
          schema: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema: { type: string }
    TooManyRequests:
      description: Rate limit exceeded (enforced by the Kong gateway, per client IP)
      headers:
        Retry-After:
          description: Seconds until the limit resets
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: API rate limit exceeded }
...
  schemas:
...
    SubmitContributionRequest:
      type: object
      required: [patch]
      properties:
        patch:
          type: object
          description: |
            Flat partial object of proposed Work field overrides, validated
            at submit time (400 with a descriptive message otherwise).
            Allowed keys: `title`, `subtitle`, `authors`, `description`,
            `language`, `isbn`, `openlibrary_id` (strings; `title` must be
            non-empty; length-capped) and `publication_year` (integer,
            -5000..current year + 5). Unknown keys, wrong types, and an
            empty patch are rejected.
          additionalProperties: true
    Contribution:
      type: object
      properties:
        id:              { type: string, format: uuid }
        work_id:         { type: string, format: uuid }
        contributor_id:  { type: integer }
        contributor_name: { type: string, description: "Display name of the contributor (omitted if the user no longer exists)" }
        status:          { type: string, enum: [pending, approved, rejected] }
        patch:
          type: object
          description: As submitted; whitelisted at Approve time, not at storage.
          additionalProperties: true
        current:
          type: object
          description: |
            The target Work's present value for each field named in `patch`
            — the "old" side of the moderator-queue diff. Resolved at read
            time, so it reflects the Work as it stands now. Keys mirror
            `patch`. Omitted when empty.
          additionalProperties: true
        reviewer_id:     { type: [integer, "null"] }
        decided_at:      { type: [string, "null"], format: date-time }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
...